Refactor e2e framework to unify eks-a API updates
Stop relying on storing the yaml content of the manifest and just use a
cluster.Config. We will follow up with more changes, simplifying how
tests update the cluster config, moving towards using
test.UpdateCLusterConfig. However, this requires changing the interfave
to the framework and hence the tests themselves. We will do that
separately and piece-meal.