Fix workload cluster Flux tests (#4381)
Since now the cluster config is not stored in yaml form but as a
cluster.Config, we need to read the cluster config file we pull from the
repo. This one is generated by the CLI and contains fields that are
necessary for a correct Flux operation, like defining a namespace for
all resources. Moreover, this ensures we only modify what is explicitly
defined in the tests as upgrade options.