Ignore kubeadm manifest preflight for ubuntu snow
Kubeadm checks if /etc/kubernetes/manifests is empty when running
kubeadm join. However, we add the kube-vip static pod manifest there in
control plane nodes, which makes the join command to fail. Bottlerocket
runs kubeadm in a different way, so it doesn't face this issue. Other
providers use useExperimentalRetryJoin, which runs kubeadm from a bash
script which adds a flag to ignore this same preflight check.