Replaces *testing.T with framework.T implementations
The *testing.T provided in tests is handy, but it panics when used outside of
a running unit test's goroutine. Replacing it with an interface allows
alternative implementations that can provide select logging and utility
without having to be executed from within a test's goroutine. This is useful
for setting up or tearing down test scaffolding (like reusable k8s clusters
for E2E tests) before or after a test execution.