Update eksa objects in cluster for skipped upgrade
This makes sure that even when a rolling upgrade is not needed, the
eks-a objects reflect the latest state driven by the CLI. This includes
new Bundles.

Before this, if there was a new Bundles that updated dependencies like
CAPI or eks-a controller but didn't require a rolling upgrade, we didn't
create the new Bundles in the cluster or update the cluster object to
reference it. It's quite an edge case since most new Bundles come with
new eks-d releases, which do trigger a rolling upgrade, so this probably
never happened in prod.