Restrict the curated packages IAM example further (#4862)
The current IAM policy example for curated packages is too open. It
would allow the IAM User to pull from any repository anywhere, including
any repository in the account. The intent is for this user to pull
curated packages from the curated packages repository. This change in
the policy example reflects that.

Note that the `ecr:GetAuthorizationToken` action does not have a resource,
so its resource must be '*'.