Set user-agent in all file readers (#4671)
This should fix the "eks-a-unknown" user-agent in requests for the
release and bundles manifests.

There is still cleanup due in the e2e framework where too many readers
are instantiated. The cluster.Spec also does too many things, we should
remove the read/load functionality it has and make it a pure data
struct. However, that would blow up the scope of this change, so leaving
that for other PRs.