Cleanup cluster.Spec
This addresses one of the standing problems with cluster.Spec's scope:
it acts both as a data object and a "manifest" reader (understand
manifest as in a yaml document read from a local file or HTTP). Being
the goal to make the Spec a pure data object, we are removing these
capabilities while maintaining the same overall program functionality:

* The Spec's ability to read manifests. This is delegated to a file
reader that any caller in need of manifests will need to accept as a
dependency.
* The Spec's ability to "build itself". Although the construction was
invoked through a constructor, the Spec's ability to read manifests was
use to build itself. This made the problem even worse, since the CLI
configuration to build the Spec (Bundles URL, etc.) bled in to the Spec
internals. Now this is all done through a FileSpecBuilder. For now this
lives in the `cluster` package. Once the `cli` package is available,
that would probably be a way better place since this builder logic is
exclusive to the CLI flows.