set a default region for PackageControllerClient
As the region setting is optional, it can't be assumed to be set, as was
happening in places.

Since it seems safe to have a default value set, and there was a default value
already defined, I've gone ahead and set it.

Now things can assume the value of the eksaRegion field is a non-zero value.