Improve envtest obj creation (#5175)
For objects without a status or with an empty status, we weren't waiting
for them to be ready. So depending on the list of objects, it was
possible for CreateObjs to return before all objects were available
through the client.

Now for objects that don't require an status update, we skip the status
update and just wait for the object to be readable.