change signaling method for FCL curated packages support (#5315)
Previously, the creation a of package bundle controller was executed via a
helm chart release. Now that helm chart will create a package instead. The end
result being the same, that the packages controller will notice the newly
created resource and do what it needs to support curated packages on the new
workload cluster.

One additional value is passed to the helm invocation to help the packages
controller take the appropriate actions.