Use kubectl for kube-proxy upgrader calls
Before the kube-proxy upgrader
(https://github.com/aws/eks-anywhere/pull/5345) all API calls to the
management cluster came either from kubectl or clusterctl, which both
happened to be run in a docker container in the admin machine. This was
the first piece of code that introduced the use of Kubernetes Go client
directly from the CLI binary. This means that if a user was relying on
this internal implementation (explicit interface vs implicit interface),
their system could break if it wasn't setup to give the CLI network
connectivity to the kind cluster.

This PR "reverts" the addition of that new paradigm byt changing the
underlying client implementation to use kubectl commands.