use --password-stdin to prevent process table leaking
Modifies the `helm registry login` command to use the `--password-stdin` flag
which prevents leaking the password on the process table.

This might also be the cause of a bug: https://github.com/aws/eks-anywhere-packages/issues/923

Though as far as I can tell, the helm registry login command still accepts the
`--password flag`, though it prints a warning.