Remove last-applied annotation for kubectl replace
Even if --save-config=false is set (which is the default), kubectl
replace will not only respect the last-applied annotation if present in
the object, but it will update it with the provided state of the
resource. This includes the metadata.resourceVersion. This breaks future
uses of kubectl apply. Since those commands' input never provide the
resourceVersion, kubectl will send a request trying to remove that
field. That is obviously not a valid request, so it gets rejected by the
kube API server. To avoid this, we simply remove the annotation when
passing it to the replace command.

Important to not that it's not recommended to use both imperative and
"declarative" commands for the same resource. Unfortunately our CLI
makes extensive use of client side apply. Although not ideal, this
change allows us to perform updates (using replace) where idempotency is
necessary while maintaining the ability to continue to use apply.