Use controller direct access to management cluster
By injecting the pod metadata as env vars in the controller, we allow
ClusterCacheTracker to detect that a management cluster is self
managed. Then it uses direct access to the control plane (through the
host in the incluster config) instead of going through the host
defined in the capi kubeconfig, which in our case goes through the
kube-vip load balancer.

More info: https://github.com/kubernetes-sigs/cluster-api/pull/6836