Patch incorrect CloudStack upgrade comparison (#5910)
* Patch incorrect CloudStack upgrade comparison

A subset of fields for the EKSA CloudStack API do not trigger creation of new
CAPI templates when they are removed. The issue stems from the
DeepDerivative function used to compare old with new to identify a diff.
It does not compare fields when their type is defaulted meaning removal
of the data alltogether isn't considered a change.

This change compares the old and new objects in both directions ensuring
we identify a diff.

* Rename CloudStack unit test