Add setting webhook data via env and cli flags:
This enables the use of the Rufio webhook provider.
The webhook provider needs a secret and a URL defined.
All BMC machines will use this one secret.

As part of enabling the webhook secret to be captured
by the CLI as an environment variable, all CLI flags are
now accessible as env vars. The env var for each flag is
the flag name but in all uppercase and the dash "-" symbol
converted to an underscore "_".

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>