Patch TinkerbellMachineTemplate omission
TinkerbellMachineTemplates are immutable. When we scale a Tinkerbell
cluster our algorithm omits them from the generated CAPI objects. The
algorithm split objects by ---. When a cluster spec is configured
with a PEM encoded secret this can cause incorrect splitting of YAML
documents.

This change ensures the YAML file is split correctly into documents.