Add workflow for management kindless upgrades
This branchs out the code at the command level and uses a different
workflow if kindless upgrades are enabled. This facilitates the
transition once we are ready to flip the switch and avoids having to
maintain multiple codepaths in the existing workflow, which from
experience, it has proven to be difficult and prone to errors. The
tradeoff is we need to mainetain the same functionality in two workflows
until we decide to deprecate the old one.