Separate etcd upgrade changes from kcp changes (#6496)
In order to have more control over the upgrade process and make sure
than in case of error the cleanup is easier, we upgrade etcd separately
from the CP nodes for external etcd topologies. The controller will
apply etcd changes in isolation and wait for etcd to be ready before
moving on with the CP nodes.