Use bmclib directly: (#6851)
* Use bmclib directly:

Going through Rufio just for the bmclib
client means that we had to take on Rufio's
dependencies. This is problematic because Rufio
uses the controller-runtime lib. Using bmclib directly
alleviates us from this concern.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Add Rufio CRDs for envtest:

This avoids having to pull in Rufio as
a dependency.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Update tests that were comparing logr.Logger:

New version of logr.Logger broken some tests.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

---------

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>