Update aws sdk modules to latest version (#7415)
This should resolve an issue with incompatible aws sdk modules. This is
due to a recent update we did where we bumped some of them but not all.
Normally this should be fine, however, a change in the aws sdk made old
modules incompatible with new modules.

More info: https://github.com/aws/aws-sdk-go-v2/issues/2370#issuecomment-1814903382