Use real version for dev releases (#7463)
Dev releases are a bit special: we generate new them all the time, very
fast. For this reason, we don't use a simple major.minor.patch semver,
but we include build metadata. In particular we use
v{major}.{minor}.{patch}-dev+build.{number} with number being a
monotonically increasing integer that is bumped every time a new dev
release is built.

Until now we always used v0.0.0-dev+something. This PR changes the
versioning. The version we use now depends on the HEAD: main vs release
branches:
* For main, we use the next minor version to the latest tag available.
  For example, if the latest prod release is v0.18.5, the version used
  for dev releases will be v0.19.0-dev+build.{number}. This aligns with
    the fact that the code in main belongs to the next future prod
    release v0.19.0.
* For release-* branches, we use the next patch version to the latest
  available tag for that minor version. For example, for release-0.17,
  if the latest latest prod release is for v0.17 is v0.17.7, dev
    releases will follow v0.17.8-dev+build.{number}.

In order to avoid the dev Release manifest growing forever, we trim the
included releases to a max size, dropping always the oldest one. Take
this in mind if using a particular version locally. If you do it for too
long, it might become unavailable. If it does, just rebuild your CLI.