Clean up funcs around power on and off:
Removed skipped tinkerbell tests that used power
on or off. Removed power on and off from all other
tests. We should not be powering on or off machines
manually. ValidateHardwareDecommissioned will now
validate machines are off and then if any are not
off we call powerOffHardware and fail the test. We
power off so that subsequent runs won't have issues
with multiple Boots DHCP servers running.

Also updated the logic around checking the response
from bmclib for power state. Depending on the provider
the power state comes back different. For example, with
ipmitool the result will be "Chassis Power is off", while
with redfish the result will be "off".

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>