Allow to set tests to run in quick e2e buildspec (#7553)
This allows to manually run a pipeline with this buildspec and running
only one or a particular subset of tests.

It introduces the value "QUICK" that allows to run all configured tests
for the quick pipeline without knowing the tests. This allows a manual
pipeline to run all quick tests without passing an empty variable, which
is not allowed.