Remove skipped test file:
Go already has a construct for skipping tests:
`t.Skip`. This file based method to skip tests
is unnecessary, adds unnecessary code, and is
yet another piece of domain knowledge that is not
documentated anywhere and difficult to understand
its scope and use. By adding `t.Skip` to all tests
that need to be skipped we eliminate all of these
concerns.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>