Fix cluster direcftory being created with root ownership
When hostpaths used for bind mounting in Docker don't exist at the time
the container is created, dockerd creates the directory. dockerd runs as
root on most systems hence the directory is created with root ownership.

This ensures the directory exists before we attempt to launch the tools
container that bind mounts the cluster directory.