Group all dependabot go changes in one PR a week
Having one PR per module can be annoying. On top of that, a lot of the
times, changes in main have effects in the release module (since we do
an import with a replace from disk).

The drawback of this change is that when code changes are required for
a particular module, the go.mod change will be mixed with other modules,
thus requiring us to separate the problematic node from the rest before
merging. From the experience running this bot during the past few
months, there is usually no work to do except to approve, no code
required most of the time. Hence this will reduce the amount of PRs
we need to review and the amount of time we need to manually update them
to simply run go mod tidy.