Update CLI logging output to be cluster type aware: (#8231)
* Update CLI logging output to be cluster type aware:

We have been leaking implementation details around
how we create clusters. While CAPI calls the initial
cluster a workload cluster and then pivots its control plane
objects to make it a management cluster this is an
implementation detail that we shouldn't be exposing to
users.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Fix test

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Use existing method to check if cluster is management

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Hardcode log lines for management cluster create:

This func is only for management cluster anyway,
so there is no need to check if its a management cluster
or not.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

---------

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>