Fix "Committer identity unknown" when patching: (#8951)
* Fix "Committer identity unknown" when patching:

git am is failing to apply the theme patch
because amplify doesn't have a git user and email
configured. This configures a user and email so
the git am doesn't fail.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Fail the make goal if the patch doesn't work:

This will make sure that a patch failure doesnt
fail silently.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* git patch with placeholder user and email:

This allows the patch to always apply. This
is needed for CI environments that don't have
git user and email setup. This won't affect
any global git config locally and doesn't affect
the patch either.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

---------

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>