Enable hardware Provisioning through ISO booting for baremetal Provider
Upstream CAPT/Tink recently added support for booting the hardware
through ISO. This change adds the necessary changes to our cluster spec
notably TinkerbellDataCenterConfig to be able to provision the hardware
using an ISO. ISO booting removes the dependency of the admin machine to
provision the cluster to be in the same L2 as the hardware. Additionally
offers static IPAM. The PR adds necessary changes required in our
TinkerbellmachineTemplate to be able to ISO boot. The default behavior
would still be to netboot. PR also adds required changes at the time of
Tinkerbell stack installation so that Smee has all the required flags
set to be able to serve ISO and handles toggling between the bootstrap
IP and the actual TinkerbelIP.

Signed-off-by: Rahul Ganesh <rahulgab@amazon.com>