Set LicenseToken field only when version is greater than v0.22 (#9264)
LicenseToken field on cluster spec will be introduced in 'v0.22.0'.
At present since we set this field when initialize the end-to-end
test object it gets set even on the previous minor versions where
it's not intended to. This change sets this field only on main, or
if the version is higher than 'v0.22.0'. This change also cleans up
some old cleanup code in our e2e framework that was used only during
way old releases such as 'v0.6.1' and is no longer needed.

Signed-off-by: Rahul Ganesh <rahulgab@amazon.com>
Co-authored-by: EKS Distro PR Bot <aws-model-rocket-bots+eksdistroprbot@amazon.com>