Add skip-validations flag to upgrade management components command (#9481)
Upgrade cluster command already supports a skip-validations flag that lets users
skip a preflight that's skippable. Add similar functionality to upgrade management-components
command. The only preflight that will be skippable in this command would be the 'eksa-version-skew'
which checks for version skew during an upgrade is not more than 1 minor version. Also for the upgrade
command controller wasn't honoring the skipped validations passed through the flag. This change also fixes
the controller to respect the flag.

Signed-off-by: Rahul Ganesh <rahulgab@amazon.com>
Co-authored-by: Rahul Ganesh <rahulgab@amazon.com>