Remove --insecure-skip-tls-verify flag from Helm uninstall command
When moving from the legacy tinkerbell chart to upstream chart we
uninstall the legacy chart. For some configurations such as airgapped
with registry mirror configurations where tls is not setup this might
end up breaking the upgrade workflow when passing in the '--insecure-skip-tls-verify'
flag. Remove the flag from helm uninstall command as it is not supported.

Signed-off-by: Rahul Ganesh <rahulgab@amazon.com>